backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]
